Fix dark mode issues
