+++ src/app/experimental/fine-tune/page.tsx
@@ -4,7 +4,7 @@
 import * as React from 'react';
 import '@patternfly/react-core/dist/styles/base.css';
 import { AppLayout, FeaturePages } from '@/components/AppLayout';
-import FineTuning from '@/components/Experimental/FineTuning';
+import FineTuning from '@/components/Experimental/FineTuning/FineTuningJobs';
 
 const FineTune: React.FunctionComponent = () => {
   return (
